requirements.txt
hyperliquid-python-sdk==0.4.0
ccxt==4.1.50
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.11.0